+++ src/semantic/checks/block.rs
@@ -1,0 +1,17 @@
+use crate::{
+    error_handling::MistiError,
+    semantic::{impls::SemanticCheck, symbol_table::SymbolTable},
+    syntax::ast::BlockMember,
+};
+
+impl<'a> SemanticCheck for BlockMember<'a> {
+    // TODO: A block may contain a function declaration statement,
+    // but (afaik) those are not allowed inside conditionals/loops
+    // somehow detect those?
+    fn check_semantics(&self, scope: &SymbolTable) -> Result<(), MistiError> {
+        match self {
+            BlockMember::Stmt(s) => s.check_semantics(scope),
+            BlockMember::Expr(e) => e.check_semantics(scope),
+        }
+    }
+}