+++ src/semantic/checks/function_declaration.rs
@@ -1,7 +1,7 @@
 use crate::{
     error_handling::{semantic_error::SemanticError, MistiError},
     semantic::{impls::SemanticCheck, symbol_table::SymbolTable, types::Type},
-    syntax::ast::{BlockMember, FunctionDeclaration, Statement},
+    syntax::ast::FunctionDeclaration,
 };
 
 impl SemanticCheck for FunctionDeclaration<'_> {
@@ -31,27 +31,7 @@ impl SemanticCheck for FunctionDeclaration<'_> {
         // TODO: Check the return type of the function body
         // This should be the last expression in the block
         for stmt in self.block.members.iter() {
-            match stmt {
-                BlockMember::Stmt(Statement::Binding(b)) => {
-                    if let Err(err) = b.check_semantics(&function_scope) {
-                        return Err(err);
-                    }
-                }
-                BlockMember::Stmt(Statement::FnDecl(f)) => {
-                    // TODO: (for now) a function cannot be declared inside another function.
-                    let error = SemanticError {
-                        error_start: f.identifier.position,
-                        error_end: f.identifier.get_end_position(),
-                        reason: format!(
-                            "A function cannot be defined inside another function."
-                        ),
-                    };
-
-                    return Err(MistiError::Semantic(error));
-                }
-                BlockMember::Stmt(Statement::Conditional(_)) => unimplemented!("check conditional"),
-                BlockMember::Expr(e) => e.check_semantics(scope)?,
-            }
+            stmt.check_semantics(&function_scope)?;
         }
 
         // TODO: Check that the return type of the function