+++ src/semantic/checks/mod.rs
@@ -1,4 +1,5 @@
 pub mod binding;
+pub mod block;
 pub mod expression;
 pub mod function_declaration;
 pub mod top_level_declaration;