+++ src/syntax/parsers/mod.rs
@@ -1,7 +1,8 @@
 pub mod binding;
 pub mod block;
+pub mod conditional;
 pub mod expression;
+pub mod for_loop;
 pub mod function_declaration;
 pub mod module;
 pub mod statement;
-pub mod conditional;